Send more info to logs

+++ src/main/ManagedChildProcess.ts
@@ -5,16 +5,25 @@ import { EventEmitter } from "events";
  * A Child Process which automatically logs all output to the console
  */
 export default class ManagedChildProcess extends EventEmitter {
-  private process: ChildProcess;
+  private process?: ChildProcess;
 
   constructor(
     private name: string,
-    command: string,
-    args: string[],
-    cwd: string
+    private command: string,
+    private args: string[],
+    private cwd: string
   ) {
     super();
-    this.process = spawn(command, args, { cwd });
+  }
+
+  spawn() {
+    if (this.process) {
+      throw Error('Cannot spawn already spawned process');
+    }
+
+    this.process = spawn(this.command, this.args, { cwd: this.cwd });
+
+    this.emit('output', `${this.name} has been started\n`);
 
     this.process.stdout.on('data', (data: Buffer) => {
       // BUG: This is only shows after the user presses CTRL+C. It does not
@@ -22,26 +31,32 @@ export default class ManagedChildProcess extends EventEmitter {
       const output = data.toString('utf8');
       const namedOutput = this.addNameToOutput(output);
       this.emit('output', namedOutput);
-      process.stdout.write(namedOutput);
     });
 
-		this.process.stderr.on('data', (data: Buffer) => {
+    this.process.stderr.on('data', (data: Buffer) => {
       const output = data.toString('utf8');
       const namedOutput = this.addNameToOutput(output);
       this.emit('output', namedOutput);
-      process.stderr.write(namedOutput);
-		});
+    });
 
-		this.process.on('close', (code) => {
-			console.log(`${this.name} exited with code ${code}`);
-		});
+    this.process.on('exit', (code, signal) => {
+      if (code) {
+        this.emit('output', `${this.name} exited with code ${code}\n`);
+      } else {
+        this.emit('output', `${this.name} exited with signal ${signal}\n`);
+      }
+
+      this.process = undefined;
+    });
   }
 
   /**
    * Politely ask the child process to exit
    */
   kill() {
-    return this.process.kill();
+    if (this.process) {
+      this.process.kill();
+    }
   }
 
   /**
